MARS-1165 Refactor to create `ViewAttributeDialog`
